Edit: Move iCommand to an interfaces directory
