+++ src/commands/purge.ts
@@ -1,5 +1,5 @@
 import { Client, DMChannel, Message } from 'discord.js';
-import ICommand from '../library/iCommand';
+import ICommand from '../library/interfaces/iCommand';
 
 let Purge: ICommand;
