+++ src/library/interfaces/iCommand.ts
@@ -1,5 +1,5 @@
 import { Client, Message } from "discord.js";
-import Commands from "./commands";
+import Commands from "../commands";
 
 /**
  * An interface for all commands to extend, representing the API that all
